test: split autonomous capture coverage
