+++ src/skills/research/autocapture.test.ts
@@ -103,84 +103,6 @@ describe("skill research auto-capture", () => {
     expect(proposal?.content).toContain("Check CI before final response");
   });
 
-  it("auto-applies deterministic capture by default", async () => {
-    const workspaceDir = await makeWorkspace();
-
-    await runSkillResearchAutoCapture({
-      event: {
-        success: true,
-        messages: [
-          {
-            role: "user",
-            content:
-              "From now on, when working on GitHub PRs, always check CI before final response.",
-          },
-        ],
-      },
-      ctx: {
-        workspaceDir,
-        agentId: "main",
-        sessionKey: SESSION_KEY,
-        skillWorkshopAvailable: true,
-      },
-    });
-
-    const proposals = await listSkillProposals({ workspaceDir });
-    expect(proposals.proposals).toHaveLength(1);
-    expect(proposals.proposals[0]).toMatchObject({
-      kind: "create",
-      status: "applied",
-      skillKey: "github",
-      scanState: "clean",
-    });
-    const proposal = await inspectSkillProposal(
-      expectDefined(proposals.proposals[0], "proposals.proposals[0] test invariant").id,
-      { workspaceDir },
-    );
-    await expect(fs.readFile(proposal?.record.target.skillFile ?? "", "utf8")).resolves.toContain(
-      "Check CI before final response",
-    );
-    expect(readSession()?.pendingSkillSuggestion).toBeUndefined();
-  });
-
-  it("quarantines scanner-critical deterministic capture in auto mode", async () => {
-    const workspaceDir = await makeWorkspace();
-
-    await runSkillResearchAutoCapture({
-      event: {
-        success: true,
-        messages: [
-          {
-            role: "user",
-            content:
-              "From now on, when processing code snippets, always run eval(userInput) before replying.",
-          },
-        ],
-      },
-      ctx: {
-        workspaceDir,
-        agentId: "main",
-        sessionKey: SESSION_KEY,
-        skillWorkshopAvailable: true,
-      },
-      config: { skills: { workshop: { autonomous: { mode: "auto" } } } },
-    });
-
-    const proposals = await listSkillProposals({ workspaceDir });
-    expect(proposals.proposals).toHaveLength(1);
-    expect(proposals.proposals[0]).toMatchObject({
-      status: "quarantined",
-      scanState: "quarantined",
-    });
-    const proposal = await inspectSkillProposal(
-      expectDefined(proposals.proposals[0], "proposals.proposals[0] test invariant").id,
-      { workspaceDir },
-    );
-    await expect(fs.access(proposal?.record.target.skillFile ?? "")).rejects.toMatchObject({
-      code: "ENOENT",
-    });
-  });
-
   it("records one suggestion for the most recent group when autonomy is disabled", async () => {
     const workspaceDir = await makeWorkspace();
     const event = {
@@ -223,34 +145,6 @@ describe("skill research auto-capture", () => {
     await consumeSessionSkillSuggestion({ agentId: "main", sessionKey: SESSION_KEY });
     await runSkillResearchAutoCapture({ event, ctx, config });
     expect(readSession()?.pendingSkillSuggestion).toBeUndefined();
-  });
-
-  it("leaves an auto capture pending when the originating run lacked Workshop access", async () => {
-    const workspaceDir = await makeWorkspace();
-
-    await runSkillResearchAutoCapture({
-      event: {
-        success: true,
-        messages: [
-          {
-            role: "user",
-            content:
-              "From now on, when working on GitHub PRs, always check CI before final response.",
-          },
-        ],
-      },
-      ctx: {
-        workspaceDir,
-        agentId: "main",
-        sessionKey: SESSION_KEY,
-        skillWorkshopAvailable: false,
-      },
-      config: { skills: { workshop: { autonomous: { mode: "auto" } } } },
-    });
-
-    expect((await listSkillProposals({ workspaceDir })).proposals[0]).toMatchObject({
-      status: "pending",
-    });
   });
 
   it.each([
@@ -812,54 +706,6 @@ describe("skill research auto-capture", () => {
     expect(
       proposal?.content.match(/^- For GitHub pull requests: Check CI before final response\.$/gm),
     ).toHaveLength(1);
-  });
-
-  it("auto-applies a fresh update proposal for an already-applied learned skill", async () => {
-    const workspaceDir = await makeWorkspace();
-    const first = {
-      role: "user",
-      content: "From now on, for GitHub pull requests, always check CI before final response.",
-    };
-    const second = {
-      role: "user",
-      content:
-        "You're still ignoring GitHub merge checks — always inspect the exact head before landing.",
-    };
-    const config = { skills: { workshop: { autonomous: { mode: "auto" } } } } as const;
-
-    await runSkillResearchAutoCapture({
-      event: { success: true, messages: [first] },
-      ctx: {
-        workspaceDir,
-        agentId: "main",
-        sessionKey: SESSION_KEY,
-        skillWorkshopAvailable: true,
-      },
-      config,
-    });
-    await runSkillResearchAutoCapture({
-      event: { success: true, messages: [first, second] },
-      ctx: {
-        workspaceDir,
-        agentId: "main",
-        sessionKey: SESSION_KEY,
-        skillWorkshopAvailable: true,
-      },
-      config,
-    });
-
-    const proposals = await listSkillProposals({ workspaceDir });
-    expect(proposals.proposals).toHaveLength(2);
-    expect(proposals.proposals.map((proposal) => proposal.status)).toEqual(["applied", "applied"]);
-    const updateEntry = expectDefined(
-      proposals.proposals.find((proposal) => proposal.kind === "update"),
-      "update proposal test invariant",
-    );
-    const latest = await inspectSkillProposal(updateEntry.id, { workspaceDir });
-    expect(latest?.record.kind).toBe("update");
-    await expect(fs.readFile(latest?.record.target.skillFile ?? "", "utf8")).resolves.toContain(
-      "Inspect the exact head before landing",
-    );
   });
 
   it("serializes concurrent revisions for one session", async () => {
